moved alarmhandler to alarmhandler.js
